requirements.txt
anyio==3.6.1
APScheduler==3.9.1
asgiref==3.2.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
python-telegram-bot==20.0a0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
style==1.1.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
update==0.0.1
urllib3==1.26.9
aiosqlite==0.17.0
alembic==1.8.0
python-dotenv==0.20.0
prettytable==3.3.0
click==8.1.3
aioschedule==0.5.2